use timestamps rather than isoformatted dates
serialization of timestamps is much faster and does not require
python-dateutil to parse so moving to utc timestamps.

This change also centralizes the json serialzation code to a new
serde module, which, for the most part can be imported like json
and used the same.
