+++ pyproject.toml
@@ -14,7 +14,6 @@ version = "1.0.0"
 aiohttp = "^3.6.2"
 prometheus_client = "^0.7.1"
 python = "^3.6"
-python-dateutil = "^2.8.1"
 
 [tool.poetry.dev-dependencies]
 attrs = "^19.3.0"
